Sync featuredImage between post content, settings and remote
Featured image created within the content will get synced.
Once it's added then, it will be shown in settings for confirmation.
If post is still draft, then it will be saved locally for future upload.
If it's published, then settings will also reflect this.

TODO:
Featured image created within the setting.
Once the post is published, then the settings does NOT reflect this condition.